Powell's minimization.
It gets stuck in local minima, which we can prove by setting the initial vector to all 0s and observing that it never hits 40%, which it did happen to reach with a different initial vector. We need a more aggressive method that works better on stepwise functions. The small perturbations this uses get it nowhere, so it doesn't try larger jumps. Maybe something like simulated annealing or Bayesian optimization? Continuous methods like hill-climbing or gradient descent will not avail us.