Add one of my hand-rolled examples to the tuning corpus.
It is the only test currently not passing. I tried to get the annealer to come up with a coefficient set that would make it do what I wanted while not wrecking the rest of the corpus, but I have failed so far. I hypothesize that it needs a bigger corpus or more features.