Fall back a little because I cannot defeat the linter.
eslint is grumpy that describe() and it() don't exist, and it also hates my symlinked test module existing in 2 places, since one of them is going to have nonexistent relative paths to other submodules. The fixes would be worse than the disease. If JS weren't such a horrible language, I'd remove the linter.