Allow clusters() (and distance()) to take either Nodes or Fnodes.
This lets us pass arrow functions to in-ruleset clustering calls that, for example, call .scoreFor() on fnodes to take score into account as part of the distance metric. Otherwise, we couldn't define such things inline, as the BoundRuleset wouldn't yet exist whose fnodeForElement() we would need to call.

Lift up logic about how to tell if something is a DOM element into isDomElement().