Come up with a universal math for determining rule prerequisites from a concept called type finalization, and unify its implementation. Close #58.
Now rules report only which types they finalize, and LHSs report only possible type combinations and which types are mentioned, and the unified Rule.prerequisites() implementation takes care of the rest. This greatly reduces the complexity of compound LHSs like and(), and the new Rule.prerequisites() is much easier to follow as well. It also results in an optimal query plan for `and(type('A')) -> type('A')`.

A future optimization opportunity would be to figure out prereqs at compile time, once.

* Upgrade chai to 4.x, which allegedly makes deepEqual work on Maps and Sets. I'm not so sure.
* Rule.prerequisites() now returns a set. Changed some ruleset tests accordingly.
* inwardRulesThatCouldEmit() and inwardRulesThatCouldAdd() now return empty arrays rather than undefined: one fewer special case.